Updating initial temperature conditions at itt=1
Previously, the initial surface temperatures (Tsurface) and the dampening temperature (Tdamp) were set at 30 degC for the start of the solver at the very first time step (itt=1). I have changed this now to the atmospheric temperature (Tatm) at the first time step for the Tsurface and to the average Tatm for Tdamp.