Correction of average water content in first two soil layers
Owater.OwGroundSoilTot was too high in the first two soil layers as they were set to the residual water content in the impervious fraction instead of 0. The calculations in the model will not be affected as they operate on the single column level and the corrections here only update Owater.OwGroundSoilTot.