Calculate soil heat storage property based on average soil moisture
Volumetric heat capacity and conductivity of soil on the ground was previously calculated based on the soil moisture of the first soil layer. However, in shorter dry periods, this is likely leading to too little soil volumetric heat capacity. Hence, these parameters are now calculated based on the average soil moisture.